Another major overhaul to view component handling.
I realized it's more useful if a {{component}} block can take a pre-generated component, since otherwise there's no way to pre-load asynchronous data before passing to handlebars.

The addition of the {{option}} handler makes for more flexibility in defining full component data in the handlebars template without resorting to json. However, this can lead to a lot of overhead as the compiled handlebars templates are significantly larger.

Created a new RadioGroup component to further test the {{option}} handler.

Switched to using a "main.hbs.html" convention for a component's main template.